Move PODB api call and data to App and pass in as props
